Components - Move deprecated to separate folders.
