+++ src/deprecated/dynamic-data-table.js
