+++ src/deprecated/widget-number.js
