+++ src/deprecated/widget-trend.js
